Reworked FnBuilder to be more in line with the rest of the generator
